Lombok+sqlite+spring bump

+++ testArtifact.sh
@@ -1,5 +1,6 @@
 #!/bin/bash
+set -ex
 ./mvnw install
 cd archetype-test
-rm -r test-archetype
+test -f test-archetype && rm -r test-archetype
 ./testGenerate.sh
